Introoduce AspNetCorePageLoadDiagnosticListener and add StartupHook project
Also moved transaction creation based on incoming ASP.NET Core HTTP into a shared WebRequestTransactionCreator project.
