+++ src/Elastic.Apm.AspNetCore/ApmMiddlewareExtension.cs
@@ -4,6 +4,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Elastic.Apm.Api;
 using Elastic.Apm.AspNetCore.DiagnosticListener;
 using Elastic.Apm.Config;
 using Elastic.Apm.DiagnosticSource;
@@ -38,7 +41,7 @@ namespace Elastic.Apm.AspNetCore
 		/// </param>
 		/// <param name="subscribers">
 		/// Specify which diagnostic source subscribers you want to connect. The
-		/// <see cref="AspNetCoreDiagnosticsSubscriber" /> is by default enabled.
+		/// <see cref="AspNetCoreErrorDiagnosticsSubscriber" /> is by default enabled.
 		/// </param>
 		public static IApplicationBuilder UseElasticApm(
 			this IApplicationBuilder builder,
@@ -68,7 +71,8 @@ namespace Elastic.Apm.AspNetCore
 		{
 			var subs = new List<IDiagnosticsSubscriber>(subscribers ?? Array.Empty<IDiagnosticsSubscriber>())
 			{
-				new AspNetCoreDiagnosticsSubscriber()
+				new AspNetCoreErrorDiagnosticsSubscriber(),
+				new AspNetCorePageLoadDiagnosticSubscriber()
 			};
 			agent.Subscribe(subs.ToArray());
 			return builder.UseMiddleware<ApmMiddleware>(agent.Tracer, agent);
@@ -86,5 +90,23 @@ namespace Elastic.Apm.AspNetCore
 			serviceProvider.GetService(typeof(ILoggerFactory)) is ILoggerFactory loggerFactory
 				? (IApmLogger)new NetCoreLogger(loggerFactory)
 				: ConsoleLogger.Instance;
+
+		internal static void UpdateServiceInformation(Service service)
+		{
+			string version;
+			var versionQuery = AppDomain.CurrentDomain.GetAssemblies().Where(n => n.GetName().Name == "Microsoft.AspNetCore");
+			var assemblies = versionQuery as Assembly[] ?? versionQuery.ToArray();
+			if (assemblies.Any())
+				version = assemblies.First().GetName().Version.ToString();
+			else
+			{
+				versionQuery = AppDomain.CurrentDomain.GetAssemblies().Where(n => n.GetName().Name.Contains("Microsoft.AspNetCore"));
+				var enumerable = versionQuery as Assembly[] ?? versionQuery.ToArray();
+				version = enumerable.Any() ? enumerable.FirstOrDefault()?.GetName().Version.ToString() : "n/a";
+			}
+
+			service.Framework = new Framework { Name = "ASP.NET Core", Version = version };
+			service.Language = new Language { Name = "C#" }; //TODO
+		}
 	}
 }
